chore: move query file into hooks
